+++ amplify/src/main/java/com/github/stkent/amplify/prompt/DefaultLayoutPromptViewConfig.java
@@ -17,7 +17,10 @@
 package com.github.stkent.amplify.prompt;
 
 import android.content.res.TypedArray;
+import android.os.Parcel;
+import android.os.Parcelable;
 import android.support.annotation.ColorInt;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.annotation.StyleableRes;
 
@@ -25,7 +28,7 @@ import com.github.stkent.amplify.R;
 
 //@formatter:off
 @SuppressWarnings({"PMD.ExcessiveParameterList", "checkstyle:parameternumber"})
-public final class DefaultLayoutPromptViewConfig {
+public final class DefaultLayoutPromptViewConfig implements Parcelable {
 
     @ColorInt
     private static final int DEFAULT_FOREGROUND_COLOR = 0xFFFFFFFF;
@@ -276,6 +279,51 @@ public final class DefaultLayoutPromptViewConfig {
 
     }
 
+    // Parcelable
+
+    @Override
+    public int describeContents() {
+        return 0;
+    }
+
+    @Override
+    public void writeToParcel(final Parcel dest, final int flags) {
+        dest.writeValue(this.foregroundColor);
+        dest.writeValue(this.backgroundColor);
+        dest.writeValue(this.titleTextColor);
+        dest.writeValue(this.subtitleTextColor);
+        dest.writeValue(this.positiveButtonTextColor);
+        dest.writeValue(this.positiveButtonBackgroundColor);
+        dest.writeValue(this.positiveButtonBorderColor);
+        dest.writeValue(this.negativeButtonTextColor);
+        dest.writeValue(this.negativeButtonBackgroundColor);
+        dest.writeValue(this.negativeButtonBorderColor);
+    }
+
+    protected DefaultLayoutPromptViewConfig(@NonNull final Parcel in) {
+        this.foregroundColor = (Integer) in.readValue(Integer.class.getClassLoader());
+        this.backgroundColor = (Integer) in.readValue(Integer.class.getClassLoader());
+        this.titleTextColor = (Integer) in.readValue(Integer.class.getClassLoader());
+        this.subtitleTextColor = (Integer) in.readValue(Integer.class.getClassLoader());
+        this.positiveButtonTextColor = (Integer) in.readValue(Integer.class.getClassLoader());
+        this.positiveButtonBackgroundColor = (Integer) in.readValue(Integer.class.getClassLoader());
+        this.positiveButtonBorderColor = (Integer) in.readValue(Integer.class.getClassLoader());
+        this.negativeButtonTextColor = (Integer) in.readValue(Integer.class.getClassLoader());
+        this.negativeButtonBackgroundColor = (Integer) in.readValue(Integer.class.getClassLoader());
+        this.negativeButtonBorderColor = (Integer) in.readValue(Integer.class.getClassLoader());
+    }
+
+    public static final Parcelable.Creator<DefaultLayoutPromptViewConfig> CREATOR
+            = new Parcelable.Creator<DefaultLayoutPromptViewConfig>() {
+
+        public DefaultLayoutPromptViewConfig createFromParcel(final Parcel source) {
+            return new DefaultLayoutPromptViewConfig(source);
+        }
+
+        public DefaultLayoutPromptViewConfig[] newArray(final int size) {
+            return new DefaultLayoutPromptViewConfig[size];
+        }
+    };
+
 }
-//CHECKSTYLE.ON: ParameterNumberCheck
 //@formatter:on