+++ testapp/src/main/java/com/github/stkent/testapp/ExampleApplication.java
@@ -2,6 +2,7 @@ package com.github.stkent.testapp;
 
 import android.app.Application;
 
+import com.github.stkent.amplify.Logger;
 import com.github.stkent.amplify.tracking.Amplify;
 
 public class ExampleApplication extends Application {
@@ -10,7 +11,8 @@ public class ExampleApplication extends Application {
     public void onCreate() {
         super.onCreate();
 
-        Amplify.get(this)
+        Amplify.initSharedInstance(this)
+               .setLogLevel(Logger.LogLevel.DEBUG)
                .setFeedbackEmailAddress("someone@example.com")
                .setAlwaysShow(true);
     }